Revert "LB implementation"
