+++ phish-patrol-frontend/src/app.html
@@ -4,14 +4,7 @@
 		<meta charset="utf-8" />
 		<link rel="icon" href="%sveltekit.assets%/favicon.png" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
-
-		<!---Imported Google fonts--->
-		<link rel="preconnect" href="https://fonts.googleapis.com">
-		<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-		<link href="https://fonts.googleapis.com/css2?family=Bagel+Fat+One&family=Bungee+Tint&family=Montserrat:ital,wght@0,100..900;1,100..900&family=Orbitron:wght@400..900&family=WDXL+Lubrifont+TC&display=swap" rel="stylesheet">
-		
 		%sveltekit.head%
-
 	</head>
 	<body data-sveltekit-preload-data="hover">
 		<div style="display: contents">%sveltekit.body%</div>
