--- phish-patrol-frontend/src/lib/server/scenarioStore.ts
@@ -1,17 +1,0 @@
-import type { Scenario } from '$lib/types';
-
-export const scenarios: Scenario[] = [];
-
-let currentId = 1;
-
-/**
- * Adds a new scenario generated by the AI to our in-memory list.
- */
-export function addScenario(scenarioData: Omit<Scenario, 'id'>): void {
-  const newScenario: Scenario = {
-    ...scenarioData,
-    id: currentId++,
-  };
-  scenarios.push(newScenario);
-  console.log(`[Scenario Store] Added new scenario. Total: ${scenarios.length}`);
-}
