--- phish-patrol-frontend/src/lib/stores/auth.ts
@@ -1,4 +1,0 @@
-import { writable } from 'svelte/store';
-
-export const tokenStore = writable('');
-export const visibleStore = writable(true); // control modal visibility
