--- phish-patrol/src/main/java/com/phish_patrol/config/README.md
@@ -1,1 +1,0 @@
-Contains configuration classes.
