--- phish-patrol/src/main/java/com/phish_patrol/controller/README.md
@@ -1,1 +1,0 @@
-Handles frontend requests and routes them to the appropriate service logic.
