--- phish-patrol/src/main/java/com/phish_patrol/dto/JwtResponse.java
@@ -1,17 +1,0 @@
-package com.phish_patrol.dto;
-
-public class JwtResponse {
-    private String token;
-
-    public JwtResponse(String token) {
-        this.token = token;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-}
