--- phish-patrol/src/main/java/com/phish_patrol/dto/LoginRequest.java
@@ -1,12 +1,0 @@
-package com.phish_patrol.dto;
-
-public class LoginRequest {
-    private String username;
-    private String password;
-
-    public String getUsername() { return username; }
-    public String getPassword() { return password; }
-
-    public void setUsername(String username) { this.username = username; }
-    public void setPassword(String password) { this.password = password; }
-}
