--- phish-patrol/src/main/java/com/phish_patrol/dto/README.md
@@ -1,1 +1,0 @@
-For frontend/backend communication.
