--- phish-patrol/src/main/java/com/phish_patrol/exception/README.md
@@ -1,1 +1,0 @@
-Used to organize custom exception classes.
