--- phish-patrol/src/main/java/com/phish_patrol/model/README.md
@@ -1,1 +1,0 @@
-Domain entities, data models
