--- phish-patrol/src/main/java/com/phish_patrol/service/CustomUserDetailsService.java
@@ -1,29 +1,0 @@
-package com.phish_patrol.service;
-
-import com.phish_patrol.model.User;
-import com.phish_patrol.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.*;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class CustomUserDetailsService implements UserDetailsService {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByUsername(username)
-            .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-
-        return new org.springframework.security.core.userdetails.User(
-            user.getUsername(),
-            user.getPasswordHash(),
-            List.of(new SimpleGrantedAuthority("ROLE_USER"))
-        );
-    }
-}
