--- phish-patrol/src/main/java/com/phish_patrol/service/README.md
@@ -1,1 +1,0 @@
-Core game logic
