--- phish-patrol/src/main/java/com/phish_patrol/util/README.md
@@ -1,2 +1,0 @@
-Miscellaneous helper classes go here.
-Examples- constants, conversion utilities, validation helpers, string manipulation helpers, etc.
