+++ phish-patrol/src/main/resources/application.properties
@@ -1,3 +1,1 @@
 spring.application.name=phish-patrol
-spring.data.mongodb.uri=mongodb+srv://admin:newPassword83248823@cluster0.98hagfp.mongodb.net/MyGameDB?appName=Cluster0
-spring.main.web-application-type=SERVLET
